+++ test/chart/animated_line_chart_test.dart
@@ -3,7 +3,6 @@ import 'package:fl_animated_linechart/common/pair.dart';
 import 'package:fl_animated_linechart/fl_animated_linechart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'dart:ui' as ui;
 
 import '../testHelpers/widget_test_helper.dart';
 
@@ -566,9 +565,6 @@ void main() {
   });
 
   testWidgets('Test legends showing', (WidgetTester tester) async {
-    final TestWidgetsFlutterBinding binding =
-        TestWidgetsFlutterBinding.ensureInitialized();
-
     DateTime start = DateTime.parse('2012-02-27 13:27:00');
 
     List<Map<DateTime, double>> series = [];
@@ -589,9 +585,8 @@ void main() {
       TextStyle(
           color: Colors.grey[800], fontSize: 11.0, fontWeight: FontWeight.w200),
     );
-
-    binding.window.physicalSizeTestValue = Size(400, 800);
-    binding.window.devicePixelRatioTestValue = 1.0;
+    tester.view.physicalSize = Size(400, 800);
+    tester.view.devicePixelRatio = 1.0;
 
     await tester.pumpWidget(
       buildTestableWidget(
@@ -618,7 +613,7 @@ void main() {
     await expectLater(find.byType(AnimatedLineChart),
         matchesGoldenFile('animatedLineChartWithLegends.png'));
 
-    binding.window.physicalSizeTestValue =
+    tester.view.physicalSize =
         Size(800, 600); //restting size of device screen back to default value
   });
 
@@ -748,7 +743,8 @@ void main() {
 
     await tester.pumpWidget(
       MediaQuery(
-        data: MediaQueryData.fromWindow(ui.window)
+        data: MediaQueryData.fromView(
+                WidgetsBinding.instance.platformDispatcher.views.single)
             .copyWith(size: const Size(600.0, 800.0)),
         child: buildTestableWidget(
           AnimatedLineChart(
@@ -802,5 +798,46 @@ void main() {
 
     await expectLater(find.byType(AnimatedLineChart),
         matchesGoldenFile('animatedLineChartLandscapeModeLegends.png'));
+  });
+
+  testWidgets('Test showing date names on x-axis', (WidgetTester tester) async {
+    DateTime start = DateTime.parse('2012-02-27 13:27:00');
+
+    List<Map<DateTime, double>> series = [];
+    Map<DateTime, double> line = {};
+    line[start] = 1.2;
+    line[start.add(Duration(days: 10))] = 0.5;
+    line[start.add(Duration(days: 70))] = 1.7;
+    line[start.add(Duration(days: 100))] = 1;
+
+    series.add(line);
+
+    LineChart lineChart = LineChart.fromDateTimeMaps(
+        series, [Colors.pink], ['P'],
+        showMonthsName: true);
+
+    lineChart.initialize(
+      200,
+      100,
+      TextStyle(
+          color: Colors.grey[800], fontSize: 11.0, fontWeight: FontWeight.w200),
+    );
+
+    await tester.pumpWidget(
+      buildTestableWidget(
+        AnimatedLineChart(
+          lineChart,
+          gridColor: Colors.black54,
+          textStyle: TextStyle(fontSize: 10, color: Colors.black54),
+          toolTipColor: Colors.white,
+        ),
+      ),
+    );
+
+    await tester.pumpAndSettle(Duration(seconds: 1));
+    await tester.pump(Duration(seconds: 1));
+
+    await expectLater(find.byType(AnimatedLineChart),
+        matchesGoldenFile('animatedLineChartShowDateNamesOnXaxis.png'));
   });
 }